+++ modules/library/index/generic/src/test/java/org/constellation/metadata/index/analyzer/SimpleAnalyzerTest.java
@@ -60,7 +60,9 @@ public class SimpleAnalyzerTest extends AbstractAnalyzerTest {
 
     @BeforeClass
     public static void setUpClass() throws Exception {
-        FileUtilities.deleteDirectory(configDirectory.toPath());
+        if (configDirectory.exists() && configDirectory.isDirectory()) {
+            FileUtilities.deleteDirectory(configDirectory.toPath());
+        }
         List<Object> object = fillTestData();
         GenericIndexer indexer = new GenericIndexer(object, null, configDirectory, "", new SimpleAnalyzer(Version.LUCENE_4_9), Level.FINER, true);
         indexer.destroy();
